broke up app.js for more modules
